create/rename endpoint directories to closely match Java service
