--- src/endpoints/explainers.py
@@ -1,211 +1,0 @@
-from fastapi import APIRouter, HTTPException
-from pydantic import BaseModel
-from typing import Dict, List, Optional
-from enum import Enum
-import logging
-
-router = APIRouter()
-logger = logging.getLogger(__name__)
-
-
-class ModelConfig(BaseModel):
-    target: str
-    name: str
-    version: Optional[str] = None
-
-
-class GlobalExplanationRequest(BaseModel):
-    modelConfig: ModelConfig
-
-
-@router.post("/explainers/global/lime")
-async def global_lime_explanation(request: GlobalExplanationRequest):
-    """Compute a global LIME explanation."""
-    try:
-        logger.info(
-            f"Computing global LIME explanation for model: {request.modelConfig.name}"
-        )
-        # TODO: Implement
-    except Exception as e:
-        logger.error(f"Error computing global LIME explanation: {str(e)}")
-        raise HTTPException(
-            status_code=500, detail=f"Error computing explanation: {str(e)}"
-        )
-
-
-@router.post("/explainers/global/pdp")
-async def global_pdp_explanation(request: GlobalExplanationRequest):
-    """Compute a global PDP explanation."""
-    try:
-        logger.info(
-            f"Computing global PDP explanation for model: {request.modelConfig.name}"
-        )
-        # TODO: Implement
-        return {"status": "success", "explanation": {}}
-    except Exception as e:
-        logger.error(f"Error computing global PDP explanation: {str(e)}")
-        raise HTTPException(
-            status_code=500, detail=f"Error computing explanation: {str(e)}"
-        )
-
-
-class LimeExplainerConfig(BaseModel):
-    n_samples: int = 300
-    timeout: int = 10
-    separable_dataset_ratio: float = 0.9
-    retries: int = 3
-    adaptive_variance: bool = True
-    penalize_balance_sparse: bool = True
-    proximity_filter: bool = True
-    proximity_threshold: float = 0.83
-    proximity_kernel_width: float = 0.5
-    encoding_cluster_threshold: float = 0.07
-    encoding_gaussian_filter_width: float = 0.07
-    normalize_weights: bool = False
-    high_score_feature_zones: bool = True
-    feature_selection: bool = True
-    n_features: int = 10
-    track_counterfactuals: bool = False
-    use_wlr_model: bool = True
-    filter_interpretable: bool = False
-
-
-class LimeExplanationConfig(BaseModel):
-    model: ModelConfig
-    explainer: Optional[LimeExplainerConfig] = None
-
-
-class LimeExplanationRequest(BaseModel):
-    predictionId: str
-    config: LimeExplanationConfig
-
-
-@router.post("/explainers/local/lime")
-async def local_lime_explanation(request: LimeExplanationRequest):
-    """Compute a LIME explanation."""
-    try:
-        logger.info(
-            f"Computing LIME explanation for prediction: {request.predictionId}"
-        )
-        # TODO: Implement
-        return {"status": "success", "explanation": {}}
-    except Exception as e:
-        logger.error(f"Error computing LIME explanation: {str(e)}")
-        raise HTTPException(
-            status_code=500, detail=f"Error computing explanation: {str(e)}"
-        )
-
-
-class LinkType(str, Enum):
-    LOGIT = "LOGIT"
-    IDENTITY = "IDENTITY"
-
-
-class RegularizerType(str, Enum):
-    AUTO = "AUTO"
-    AIC = "AIC"
-    BIC = "BIC"
-    TOP_N_FEATURES = "TOP_N_FEATURES"
-    NONE = "NONE"
-
-
-class SHAPExplainerConfig(BaseModel):
-    n_samples: int = 300
-    timeout: int = 10
-    link: LinkType = LinkType.IDENTITY
-    regularizer: RegularizerType = RegularizerType.AUTO
-    confidence: float = 0.95
-    track_counterfactuals: bool = False
-
-
-class SHAPExplanationConfig(BaseModel):
-    model: ModelConfig
-    explainer: Optional[SHAPExplainerConfig] = None
-
-
-class SHAPExplanationRequest(BaseModel):
-    predictionId: str
-    config: SHAPExplanationConfig
-
-
-@router.post("/explainers/local/shap")
-async def local_shap_explanation(request: SHAPExplanationRequest):
-    """Compute a SHAP explanation."""
-    try:
-        logger.info(
-            f"Computing SHAP explanation for prediction: {request.predictionId}"
-        )
-        # TODO: Implement
-        return {"status": "success", "explanation": {}}
-    except Exception as e:
-        logger.error(f"Error computing SHAP explanation: {str(e)}")
-        raise HTTPException(
-            status_code=500, detail=f"Error computing explanation: {str(e)}"
-        )
-
-
-class CounterfactualExplainerConfig(BaseModel):
-    n_samples: int = 100
-
-
-class CounterfactualExplanationConfig(BaseModel):
-    model: ModelConfig
-    explainer: Optional[CounterfactualExplainerConfig] = None
-
-
-class CounterfactualExplanationRequest(BaseModel):
-    predictionId: str
-    config: CounterfactualExplanationConfig
-    goals: Optional[Dict[str, str]] = None
-    explanationConfig: Optional[CounterfactualExplanationConfig] = None
-
-
-@router.post("/explainers/local/cf")
-async def local_counterfactual_explanation(request: CounterfactualExplanationRequest):
-    """Compute a Counterfactual explanation."""
-    try:
-        logger.info(
-            f"Computing Counterfactual explanation for prediction: {request.predictionId}"
-        )
-        # TODO: Implement
-        return {"status": "success", "explanation": {}}
-    except Exception as e:
-        logger.error(f"Error computing Counterfactual explanation: {str(e)}")
-        raise HTTPException(
-            status_code=500, detail=f"Error computing explanation: {str(e)}"
-        )
-
-
-class TSSaliencyExplainerConfig(BaseModel):
-    timeout: int = 10
-    mu: float = 0.01
-    n_samples: int = 50
-    n_alpha: int = 50
-    sigma: float = 50
-    base_values: Optional[List[float]] = None
-
-
-class TSSaliencyExplanationConfig(BaseModel):
-    model: ModelConfig
-    explainer: Optional[TSSaliencyExplainerConfig] = None
-
-
-class TSSaliencyExplanationRequest(BaseModel):
-    predictionIds: List[str]
-    config: TSSaliencyExplanationConfig
-
-
-@router.post("/explainers/local/tssaliency")
-async def local_tssaliency_explanation(request: TSSaliencyExplanationRequest):
-    """Compute a TSSaliency explanation."""
-    try:
-        logger.info(
-            f"Computing TSSaliency explanation for predictions: {request.predictionIds}"
-        )
-        # TODO: Implement
-        return {"status": "success", "explanation": {}}
-    except Exception as e:
-        logger.error(f"Error computing TSSaliency explanation: {str(e)}")
-        raise HTTPException(
-            status_code=500, detail=f"Error computing explanation: {str(e)}"
-        )
