+++ src/endpoints/metrics/fairness/group/spd.py
